Fix FallingSand creating invisible blocks
Workaround for https://bugs.mojang.com/browse/MC-72248
If an EntityFallingBlock spawns inside a block of the same type, the
client will ALWAYS remove the block, whereas the server will only
remove it if ticksLived is 0. This creates invisible blocks on the
client. The imperfect workaround is to not spawn the falling block at
all if it will cause such a desync.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>